updated Projekt layout 2
